ref(tokens-interface): move the interface to the dto
